requirements.txt
blinker==1.6.3
click==8.1.7
flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lightgbm==4.1.0
MarkupSafe==2.1.3
numpy==1.24.4
pandas==2.0.3
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
werkzeug==3.0.0
xgboost==2.0.0
zipp==3.17.0
